Permissoes de Usuarios - uma parte feita

+++ app/src/main/java/com/example/appsenkaspi/AcaoEntity.kt
@@ -31,7 +31,7 @@ import com.example.appsenkaspi.Converters.StatusAcao
     ]
 )
 data class AcaoEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0L,
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val nome: String,
     val descricao: String,
     val dataInicio: Date,
